+++ src/main.js
@@ -18,6 +18,22 @@ Vue.prototype.$axios = axios.create({
   }
 })
 
+Vue.mixin({
+  methods: {
+    formatTrackTime (v) {
+      var sec_num = parseInt(v, 10);
+      var hours   = Math.floor(sec_num / 3600);
+      var minutes = Math.floor((sec_num - (hours * 3600)) / 60);
+      var seconds = sec_num - (hours * 3600) - (minutes * 60);
+
+      if (hours && hours   < 10) {hours   = "0"+hours;}
+      if (minutes < 10) {minutes = "0"+minutes;}
+      if (seconds < 10) {seconds = "0"+seconds;}
+      return (hours ? hours+':' : '')+minutes+':'+seconds;
+    }
+  }
+})
+
 Vue.config.productionTip = false
 
 new Vue({